+++ src/utils/math/Vector2D.cpp
@@ -1,19 +1,14 @@
-//
-// Created by jkirc on 8/2/2025.
-//
-
 #include "Vector2D.h"
 
 // Constructors
-Vector2D::Vector2D() : x(0), y(0) {}
-
-Vector2D::Vector2D(int x, int y) : x(x), y(y) {}
+Vector2D::Vector2D() : x(0.0), y(0.0) {}
+Vector2D::Vector2D(double x, double y) : x(x), y(y) {}
 
 // Accessors
-int Vector2D::getX() const { return x; }
-int Vector2D::getY() const { return y; }
-void Vector2D::setX(int x) { this->x = x; }
-void Vector2D::setY(int y) { this->y = y; }
+double Vector2D::getX() const { return x; }
+double Vector2D::getY() const { return y; }
+void Vector2D::setX(double x) { this->x = x; }
+void Vector2D::setY(double y) { this->y = y; }
 
 // Core operations
 Vector2D Vector2D::add(const Vector2D& other) const {
@@ -28,31 +23,31 @@ Vector2D Vector2D::multiply(const Vector2D& other) const {
     return Vector2D(x * other.x, y * other.y);
 }
 
-float Vector2D::length() const {
-    return std::sqrt(static_cast<float>(x * x + y * y));
+double Vector2D::length() const {
+    return std::sqrt(x * x + y * y);
 }
 
-float Vector2D::lengthSquared() const {
-    return static_cast<float>(x * x + y * y);
+double Vector2D::lengthSquared() const {
+    return x * x + y * y;
 }
 
 Vector2D Vector2D::normalize() const {
-    float len = length();
-    if (len == 0) return Vector2D(0, 0);
-    return Vector2D(static_cast<int>(x / len), static_cast<int>(y / len));
+    double len = length();
+    if (len == 0.0) return Vector2D(0.0, 0.0);
+    return Vector2D(x / len, y / len);
 }
 
-int Vector2D::dot(const Vector2D& other) const {
+double Vector2D::dot(const Vector2D& other) const {
     return x * other.x + y * other.y;
 }
 
-float Vector2D::distanceTo(const Vector2D& other) const {
-    int dx = x - other.x;
-    int dy = y - other.y;
-    return std::sqrt(static_cast<float>(dx * dx + dy * dy));
+double Vector2D::distanceTo(const Vector2D& other) const {
+    double dx = x - other.x;
+    double dy = y - other.y;
+    return std::sqrt(dx * dx + dy * dy);
 }
 
-// Operator overloads
+// Operators
 Vector2D Vector2D::operator+(const Vector2D& other) const {
     return add(other);
 }
@@ -73,7 +68,7 @@ bool Vector2D::operator!=(const Vector2D& other) const {
     return !(*this == other);
 }
 
-int Vector2D::operator[](int index) const {
+double Vector2D::operator[](int index) const {
     if (index == 0) return x;
     else if (index == 1) return y;
     throw std::out_of_range("Index out of range for Vector2D");
@@ -86,13 +81,13 @@ std::ostream& operator<<(std::ostream& os, const Vector2D& vec) {
 
 // Static helpers
 Vector2D Vector2D::zero() {
-    return Vector2D(0, 0);
+    return Vector2D(0.0, 0.0);
 }
 
 Vector2D Vector2D::unitX() {
-    return Vector2D(1, 0);
+    return Vector2D(1.0, 0.0);
 }
 
 Vector2D Vector2D::unitY() {
-    return Vector2D(0, 1);
+    return Vector2D(0.0, 1.0);
 }
